chore(florida): Move RequestManager to state module
